requirements/development.txt
Django==1.6.5
South==1.0
WebOb==1.4
argparse==1.2.1
bobo==2.1.0
boto==2.30.0
django-dbbackup==1.80.2
django-storages==1.1.8
django-widget-tweaks==1.3
dropbox==2.1.0
path.py==5.0
psycopg2==2.5.3
pytz==2014.4
six==1.7.3
urllib3==1.8.3
wsgiref==0.1.2
django-less==0.7.2
Markdown==2.4.1
django-crispy-forms==1.4.0
django-localflavor==1.0
icalendar==3.8.1